Moved int test to cursorpage-jpa
- removed testapp (already copied in examples)
