+++ cursorpaging-examples/maven-setup/src/main/java/io/vigier/sbcpreleasetest/repository/DataRecordRepository.java
@@ -1,7 +1,7 @@
-package io.vigier.sbcpreleasetest.repository;
+package io.vigier.cursorpaging.testapp.repository;
 
 import io.vigier.cursorpaging.jpa.repository.CursorPageRepository;
-import io.vigier.sbcpreleasetest.model.DataRecord;
+import io.vigier.cursorpaging.testapp.model.DataRecord;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
