+++ cursorpaging-jpa/src/test/java/io/vigier/cursorpaging/jpa/itest/TestApplication.java
@@ -1,4 +1,4 @@
-package io.vigier.cursorpaging.testapp;
+package io.vigier.cursorpaging.jpa.itest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
