+++ cursorpaging-jpa/src/test/java/io/vigier/cursorpaging/jpa/itest/config/JpaConfig.java
@@ -1,7 +1,7 @@
-package io.vigier.cursorpaging.testapp.config;
+package io.vigier.cursorpaging.jpa.itest.config;
 
 import io.vigier.cursorpaging.jpa.bootstrap.CursorPageRepositoryFactoryBean;
-import io.vigier.cursorpaging.testapp.TestApplication;
+import io.vigier.cursorpaging.jpa.itest.repository.DataRecordRepository;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
@@ -10,7 +10,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
  * would require an entity manager factory (bean).
  */
 @Configuration
-@EnableJpaRepositories( basePackageClasses = TestApplication.class, repositoryFactoryBeanClass = CursorPageRepositoryFactoryBean.class )
+@EnableJpaRepositories( basePackageClasses = DataRecordRepository.class,
+        repositoryFactoryBeanClass = CursorPageRepositoryFactoryBean.class )
 public class JpaConfig {
 
 }
