+++ cursorpaging-jpa/src/test/java/io/vigier/cursorpaging/jpa/itest/repository/DataRecordRepository.java
@@ -1,7 +1,7 @@
-package io.vigier.cursorpaging.testapp.repository;
+package io.vigier.cursorpaging.jpa.itest.repository;
 
+import io.vigier.cursorpaging.jpa.itest.model.DataRecord;
 import io.vigier.cursorpaging.jpa.repository.CursorPageRepository;
-import io.vigier.cursorpaging.testapp.model.DataRecord;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
