--- cursorpaging-testapp/src/main/java/io/vigier/cursorpaging/testapp/api/model/DtoDataRecord.java
@@ -1,16 +1,0 @@
-package io.vigier.cursorpaging.testapp.api.model;
-
-import java.time.OffsetDateTime;
-import lombok.Data;
-
-@Data
-public class DtoDataRecord {
-
-    private String id;
-
-    private String name;
-
-    private OffsetDateTime createdAt;
-
-    private OffsetDateTime modifiedAt;
-}
