--- cursorpaging-testapp/src/main/java/io/vigier/cursorpaging/testapp/api/model/mapper/DtoDataRecordMapper.java
@@ -1,33 +1,0 @@
-package io.vigier.cursorpaging.testapp.api.model.mapper;
-
-import io.vigier.cursorpaging.testapp.api.model.DtoDataRecord;
-import io.vigier.cursorpaging.testapp.config.MapstructConfig;
-import io.vigier.cursorpaging.testapp.model.DataRecord;
-import java.time.Instant;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
-import java.util.UUID;
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper( config = MapstructConfig.class, unmappedTargetPolicy = ReportingPolicy.ERROR, injectionStrategy = InjectionStrategy.CONSTRUCTOR )
-public abstract class DtoDataRecordMapper {
-
-    @Mapping( target = "id" )
-    public String toId( final UUID id ) {
-        return id.toString();
-    }
-
-    public OffsetDateTime toOffsetDateTime( final Instant source ) {
-        if ( source == null ) {
-            return null;
-        }
-        return source.atOffset( ZoneOffset.UTC );
-    }
-
-    @Mapping( target = "createdAt", source = "auditInfo.createdAt", dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ" )
-    @Mapping( target = "modifiedAt", source = "auditInfo.modifiedAt", dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ" )
-    public abstract DtoDataRecord toDto( DataRecord source );
-}
