--- cursorpaging-testapp/src/main/java/io/vigier/cursorpaging/testapp/config/MapstructConfig.java
@@ -1,8 +1,0 @@
-package io.vigier.cursorpaging.testapp.config;
-
-import org.mapstruct.MapperConfig;
-
-@MapperConfig( componentModel = "spring" )
-public class MapstructConfig {
-
-}
