--- cursorpaging-testapp/src/main/java/io/vigier/cursorpaging/testapp/config/WebConfig.java
@@ -1,22 +1,0 @@
-package io.vigier.cursorpaging.testapp.config;
-
-import io.vigier.cursorpaging.jpa.api.StringToBase64StringConverter;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.hateoas.config.EnableHypermediaSupport;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableHypermediaSupport( type = { EnableHypermediaSupport.HypermediaType.HAL } )
-public class WebConfig implements WebMvcConfigurer {
-
-    /**
-     * Optional configuration to auto-convert base64 containing strings into a value class
-     *
-     * @return converter
-     */
-    @Bean
-    public StringToBase64StringConverter stringToBase64StringConverter() {
-        return new StringToBase64StringConverter();
-    }
-}
