--- cursorpaging-testapp/src/main/resources/static/index.html
@@ -1,7 +1,0 @@
-<html lang="en">
-<body>
-<h2>Cursor Paging Test App</h2>
-<p>API</p>
-<a href="swagger-ui/index.html">swagger-ui</a>
-</body>
-</html>
