--- cursorpaging-testapp/src/test/resources/io/vigier/cursor/testapp/application.yml
@@ -1,9 +1,0 @@
-spring:
-  jpa:
-    hibernate.ddl-auto: create
-    database-platform: "org.hibernate.dialect.PostgreSQLDialect"
-    show-sql: false
-#  datasource:
-#    driver-class-name: "${DB_DRIVER}"
-#    username: "${DB_USER}"
-#    password: "${DB_PASSWORD}"
