+++ settings.gradle.kts
@@ -25,7 +25,7 @@ dependencyResolutionManagement {
 
 rootProject.name = "spring-cursorpaging"
 
-include("cursorpaging-jpa", "cursorpaging-jpa-api", "cursorpaging-testapp")
+include("cursorpaging-jpa", "cursorpaging-jpa-api")
 
 develocity {
     buildScan {
